requirements.txt
streamlit
pandas
numpy 
scipy
plotly.express